updated codes

+++ bench.c
@@ -1,11 +1,11 @@
-double run_code(double number)
+int run_code(int number)
 {
     if (number == 1)
         return 1;
     else {
-        double num1 = number;
-        double num2 = number;
-        for (double i = 0; i < number * 1000; i++) {
+        int num1 = number;
+        int num2 = number;
+        for (int i = 0; i < number * 1000; i++) {
             num2 *= number;
         }
         return num2 * run_code(num1 - 1);
