+++ bench.cs
@@ -2,14 +2,14 @@ using System;
 
 public class bench
 {
-    public static double RunCode(double number)
+    public static int RunCode(int number)
     {
         if (number == 1)
             return 1;
         else {
-            double num1 = number;
-            double num2 = number;
-            for (double i = 0; i < number * 1000; i++) {
+            int num1 = number;
+            int num2 = number;
+            for (int i = 0; i < number * 1000; i++) {
                 num2 *= number;
             }
             return num2 * RunCode(num1 - 1);
