+++ bench.java
@@ -1,13 +1,13 @@
 public class bench
 {
-    public static double runCode(double number)
+    public static int runCode(int number)
     {
         if (number == 1)
             return 1;
         else {
-            double num1 = number;
-            double num2 = number;
-            for (double i = 0; i < number * 1000; i++) {
+            int num1 = number;
+            int num2 = number;
+            for (int i = 0; i < number * 1000; i++) {
                 num2 *= number;
             }
             return num2 * runCode(num1 - 1);
